Refactors Blog catgories routes
