+++ app/views/pages/blog_posts/index.html.erb
@@ -1,4 +1,4 @@
-<%= render layout: "blog_layout" do %>
+<%= render layout: "pages/blog_posts/blog_layout" do %>
   <% @posts.each do |post| %>
     <div class="post-snippet mb64">
       <div class="post-title">
@@ -13,7 +13,7 @@
             <i class="ti-tag"></i>
             <span>Categorized as
               <% post.blog_categories.each do |category| %>
-                <%= link_to category.name, pages_blog_categories_path(category.name) %>
+                <%= link_to category.name, pages_blog_category_path(category.name) %>
               <% end %>
             </span>
           </li>
