+++ app/views/pages/blog_posts/show.html.erb
@@ -6,7 +6,7 @@
   <meta property="og:image"         content="<%= image_url("topTutoringFinal.png") %>" />
 <% end %>
 
-<%= render layout: "blog_layout" do %>
+<%= render layout: "pages/blog_posts/blog_layout" do %>
   <div class="post-snippet mb64">
     <article>
       <div class="post-title">
